Do not remove empty element included in blocks (#15)
* Do not remove empty element included in blocks

* Improved template marking and cleanup

Previous solution was more coarse and tagged any XML element inside
template block construct as candidate for removal when empty.  This does
not work as people often use empty paragraphes for layout, even inside a
template construct.

The PR changes the logic and ensure that we only mark elements that are
around template tags.